% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categObsPredModelParam-class.R
\docType{class}
\name{categObsPredModelParam-class}
\alias{categObsPredModelParam-class}
\title{S4 class for the parameters of a categorical observation-prediction model}
\description{
Definition of the S4 class named \code{categObsPredModelParam} for the parameters of
a categorical observation-prediction model in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{Data}}{Object of S3 class \code{StQ} with the parameters or all data necessary to
compute them.}

\item{\code{VarRoles}}{List with components \code{Units}, \code{Domains}, \code{DesignW},
\code{Regressands}, \code{Regressors} being character vectors containing the column names
according to their respective role in the model.}

\item{\code{modelFits}}{List with components...}

\item{\code{probs}}{Data.table with the probabilities from predict...}

\item{\code{Train}}{Data.table with the set of data for training the model.}

\item{\code{Test}}{Data.table with the set of data for testing the model.}
}}

\examples{
# An empty contObsPredModelParam object:
new(Class = 'categObsPredModelParam')

\dontrun{

ObsPredPar <- new(Class = 'categObsPredModelParam',
                  Data = FD,
                  VarRoles = list(Units = 'NOrden', Domains = 'GeoLoc_35._4._2.1.5._1.2.3.'))

}

}
