% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeProbs.R
\docType{methods}
\name{computeProbs}
\alias{computeProbs}
\alias{computeProbs,categObsPredModelParam-method}
\title{Compute error probabilities according to the error model}
\usage{
computeProbs(object)

\S4method{computeProbs}{categObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{categObsPredModelParam}.}
}
\value{
A \linkS4class{data.table} with the following columns:

\enumerate{

\item One per regressor in \code{formula}.

\item Regressand in \code{formula}.

\item \code{pi}: for probabilities \eqn{\pi(x)}. See (REF).

\item \code{p11}: for probabilities \eqn{p(1|1)}. See (REF).

\item \code{p01}: for probabilities \eqn{p(0|1)}. See (REF).

\item \code{p10}: for probabilities \eqn{p(1|0)}. See (REF).

\item \code{p00}: for probabilities \eqn{p(0|0)}. See (REF).

\item \code{P11}: for probabilities \eqn{P(1|1)}. See (REF).

\item \code{P01}: for probabilities \eqn{P(0|1)}. See (REF).

\item \code{P10}: for probabilities \eqn{P(1|0)}. See (REF).

\item \code{P00}: for probabilities \eqn{P(0|0)}. See (REF).

}
}
\description{
Compute error probabilities according to the error model
}
\details{
Regarding \code{formula} see \code{\link{robustlogrm}}.

The input data set \code{modelsDT} must contain the following colums:
\enumerate{

\item LHS of \code{formula}: values of the target variable.

\item \code{fit.pi}: glm model fit for the probabilities \eqn{\pi(x)}.

\item \code{fit.1}: glm model fit for the probabilities \eqn{p(\cdot|1)}.

\item \code{fit.0}: glm model fit for the probabilities \eqn{p(\cdot|0)}.

}
}
\examples{
\dontrun{
id.vars <- c('IDENTHOGAR', 'NORDEN')
formula <- 'CNO ~ CNAE1'
probs <- computeProbs(models, id.vars, formula)
}

}
