% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editPriority.R
\name{editPriority}
\alias{editPriority}
\title{Return a \linkS4class{data.table} with the priority that each unit 
has to be analyzed.}
\usage{
editPriority(
  models,
  data,
  formula,
  targetVar,
  regressors,
  designWeight,
  id.vars,
  suffix = "_ed"
)
}
\arguments{
\item{models}{A \linkS4class{data.table} with four colums named with the LHS
variable of \code{formula}, fit.pi, fit.1 and fit.0. The first column contains 
every single value of this variable in the input data set \code{data}. 
The second, third and fourth columns contain a \code{glm} object with the fit
of the model for each variable value.}

\item{data}{data frame, list or environment (or object coercible by
\code{as.data.table} to a \linkS4class{data.table}) containing the variables
in the model (see 'Details').}

\item{formula}{an object of class \code{\link[stats]{formula}} (or one that
can be coerced to that class): a symbolic description of the model to be
fitted. The details of model specification are given under ‘Details’.}

\item{targetVar}{name of the regressand (an object of class character).}

\item{regressors}{an object of class character containing the names of regressors.}

\item{designWeight}{an object of class character containing the design weight.}

\item{id.vars}{names of the identification variables for each unit in the
input data set \code{data}.}

\item{suffix}{parameter for the name of the edited version of the variable under analysis.}
}
\value{
A \linkS4class{data.table} containing the following variables:
id.vars, regressand, regressors, design weight, P01, moments, regressors and the priority
to be analyzed.
}
\description{
\code{editPriority} computes the priority for each unit in the
input data set \code{data}. The corresponding model is applied to each unit,
depending on the value of the regressand. The empirical distribution is used 
for every single value of the regressand, in order to assign a correct 
priority to units.
}
\examples{

\dontrun{
}

}
