% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitModels}
\alias{fitModels}
\alias{fitModels,categObsPredModelParam,fitParam,logical-method}
\title{Return \linkS4class{data.table} with model fits for probabilities.}
\usage{
fitModels(object, fitParam, na.as.category)

\S4method{fitModels}{categObsPredModelParam,fitParam,logical}(object, fitParam, na.as.category = FALSE)
}
\arguments{
\item{object}{Object of class \linkS4class{categObsPredModelParam}.}

\item{fitParam}{Object of class \linkS4class{fitParam}.}

\item{na.as.category}{logical with default to \code{FALSE} to deal with NA values in the model. 
If \code{TRUE} the missing values are treated as a new category.}
}
\value{
Object of class \linkS4class{categObsPredModelParam} with the slot 
\code{modelFits}
computed with the fitted models.
}
\description{
\code{fitModels} computes model fits for probabilities.
}
\examples{

\dontrun{
  fitPar <- new(Class = 'fitParam',
  edData = FFall_AS.StQ, rawData = FGall_AS.StQ, 
  selParam = list(ntreeTry=1000, stepFactor=2, improve=0.05, 
                  trace=TRUE, plot=TRUE, doBest = TRUE, 
                  ptrain = 0.8, DD = DDactu),
                  valParam = list(edEffInd = effInd, priorBin = 5, 
                  dataVal = c('Train','Test')))
                  
 ObsPredPar1 <- new(Class = 'categObsPredModelParam',
                  Data = FGall_AS.StQ,
                  VarRoles = list(Units = IDUnits,
                  Domains = character(0),
                  DesignW = DesignW,
                  Regressands = Regressands,
                  Regressors = Regressors
                  ))
                  
ObsPredPar1 <-  fitModels(ObsPredPar1, fitPar, na.as.category)
}

}
