% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitParam-class.R
\docType{class}
\name{fitParam-class}
\alias{fitParam-class}
\title{S4 class for the parameters necessary to fit a categorical observation-prediction model}
\description{
Definition of the S4 class named \code{fitParam} for the parameters needed to fit
a categorical observation-prediction model in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{edData}}{object of class \code{\link[StQ]{StQ}} with the edited and validated values.}

\item{\code{rawData}}{object of class \code{\link[StQ]{StQ}} with the raw values.}

\item{\code{selection}}{logical vector of length 1 to indicate whether a model selection is to be
performed (\code{TRUE}) or not (\code{FALSE}).}

\item{\code{formula}}{character vector with formula(s) for the model(s) to fit.}

\item{\code{selParam}}{list of parameters for the model selection when \code{selection == TRUE}}
}}

\examples{
# An empty contObsPredModelParam object:
new(Class = 'fitParam')

\dontrun{
source('R:/USIE/Proyecto_DepSel_VarQual/datos/PreparacionStQsAnalisis.R', encoding = 'UTF-8')
fitPar <- new(Class = 'fitParam', edData = ff_2011_model.StQ, rawData = fd_2011_model.StQ,
              selection = FALSE,  formula = 'Ocupacion_35.__11.1.3._ ~ ActivEcono_35.__2.1.1._',
              selParam = list())

}

}
