% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdData.R
\docType{methods}
\name{getEdData}
\alias{getEdData}
\alias{getEdData,fitParam-method}
\title{Return slot \code{edData} from the input object}
\usage{
getEdData(object)

\S4method{getEdData}{fitParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{fitParam}.}
}
\value{
Object of class \linkS4class{StQ} corresponding to the slot \code{edData} of the input
object.
}
\description{
\code{getEdData} extracts the slot \code{edData} which contains the edited data for
the construction of a categorical observation-prediction model.
}
\examples{
\dontrun{
source('R:/USIE/Proyecto_DepSel_VarQual/datos/PreparacionStQsAnalisis.R', encoding = 'UTF-8')
fitPar <- new(Class = 'fitParam', edData = ff_2011_model.StQ, rawData = fd_2011_model.StQ,
              selection = FALSE,  formula = 'Ocupacion_35.__11.1.3._ ~ ActivEcono_35.__2.1.1._',
              selParam = list())
getEdData(fitPar)
}


}

