% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFormula.R
\docType{methods}
\name{getFormula}
\alias{getFormula}
\alias{getFormula,fitParam-method}
\title{Return slot \code{formula} from the input object}
\usage{
getFormula(object)

\S4method{getFormula}{fitParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{fitParam}.}
}
\value{
Object of class \code{formula} with the formula for models.
}
\description{
\code{getFormula} extracts the slot \code{formula} of the input object.
}
\examples{
\dontrun{
source('R:/USIE/Proyecto_DepSel_VarQual/datos/PreparacionStQsAnalisis.R', encoding = 'UTF-8')
fitPar <- new(Class = 'fitParam', edData = ff_2011_model.StQ, rawData = fd_2011_model.StQ,
              selection = FALSE,  formula = 'Ocupacion_35.__11.1.3._ ~ ActivEcono_35.__2.1.1._',
              selParam = list())
getFormula(fitPar)
}

}

