% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelFits.R
\docType{methods}
\name{getModelFits}
\alias{getModelFits}
\alias{getModelFits,categObsPredModelParam-method}
\title{Return slot \code{modelFits} from the input object}
\usage{
getModelFits(object)

\S4method{getModelFits}{categObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{categObsPredModelParam}.}
}
\value{
\linkS4class{data.table} with the fitted models.
}
\description{
\code{getModelFits} extracts the slot \code{VarRoles} of the input object.
}
\examples{

}
