% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegressands.R
\docType{methods}
\name{getRegressands}
\alias{getRegressands}
\alias{getRegressands,categObsPredModelParam-method}
\title{Return the regressands for the meaurement error models from the input object}
\usage{
getRegressands(object)

\S4method{getRegressands}{categObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{categObsPredModelParam}.}
}
\value{
Character vector with the names of the regressands.
}
\description{
\code{getRegressands} extracts the component \code{Regressands} from the slot
 \code{VarRoles} of the input object.
}
\examples{
\dontrun{
source('R:/USIE/Proyecto_DepSel_VarQual/datos/PreparacionStQsAnalisis.R', encoding = 'UTF-8')
fitPar <- new(Class = 'fitParam', edData = ff_2011_model.StQ, rawData = fd_2011_model.StQ,
              selection = FALSE,  formula = 'Ocupacion_35.__11.1.3._ ~ ActivEcono_35.__2.1.1._',
              selParam = list())
getEdData(fitPar)
}


}

