% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegressors.R
\docType{methods}
\name{getRegressors}
\alias{getRegressors}
\alias{getRegressors,categObsPredModelParam-method}
\title{Return the regressors for the meaurement error models from the input object}
\usage{
getRegressors(object)

\S4method{getRegressors}{categObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{categObsPredModelParam}.}
}
\value{
Character vector with the names of the regressors.
}
\description{
\code{getRegressors} extracts the component \code{Regressors} from the slot
 \code{VarRoles} of the input object.
}
\examples{
\dontrun{
source('R:/USIE/Proyecto_DepSel_VarQual/datos/PreparacionStQsAnalisis.R', encoding = 'UTF-8')
fitPar <- new(Class = 'fitParam', edData = ff_2011_model.StQ, rawData = fd_2011_model.StQ,
              selection = FALSE,  formula = 'Ocupacion_35.__11.1.3._ ~ ActivEcono_35.__2.1.1._',
              selParam = list())
getEdData(fitPar)
}


}
