% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelection.R
\name{getSelection}
\alias{getSelection}
\alias{getSelection,fitParam-method}
\title{Return slot \code{selection} from the input object}
\usage{
getSelection(object)

\S4method{getSelection}{fitParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{fitParam}.}
}
\value{
Logical indicating whether a model selection is requested or not in the parameters for
the fit of the model.
}
\description{
\code{getSelection} extracts the slot \code{selection} of the input object.
}
\examples{
\dontrun{
source('R:/USIE/Proyecto_DepSel_VarQual/datos/PreparacionStQsAnalisis.R', encoding = 'UTF-8')
fitPar <- new(Class = 'fitParam', edData = ff_2011_model.StQ, rawData = fd_2011_model.StQ,
              selection = FALSE,  formula = 'Ocupacion_35.__11.1.3._ ~ ActivEcono_35.__2.1.1._',
              selParam = list())
getSelection(fitPar)
}

}
