% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{Models}
\alias{Models}
\title{Return \linkS4class{data.table} with model fits for prior probabilities}
\usage{
Models(formula, data, id.vars, targetValue = "", maxit, suffix = "_ed")
}
\arguments{
\item{formula}{}

\item{data}{}

\item{id.vars}{}

\item{targetValue}{}

\item{maxit}{}

\item{suffix}{}
}
\value{
\linkS4class{data.table} with ...
}
\description{
\code{Models} computes ...
}
\examples{

\dontrun{
}

}
