% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regfit2.R
\name{regfit2}
\alias{regfit2}
\title{Fit robust logistic regression models to values of an input variable}
\usage{
regfit2(formula, data, id.vars, targetValue = "", cond, suffix = "_ed", ...)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} (or one that
can be coerced to that class): a symbolic description of the model to be
fitted. The details of model specification are given under ‘Details’.}

\item{data}{data frame, list or environment (or object coercible by
\code{as.data.table} to a \linkS4class{data.table}) containing the variables
in the model (see 'Details').}

\item{id.vars}{names of the identification variables for each unit in the
input data set \code{data}.}

\item{cond}{parameter indicating the model to adjust according to:
\enumerate{
  \item \code{''} for the \eqn{\pi} model (marginal probability for the
  target variable value).
  \item \code{'0'} (or \code{0} or \code{FALSE}) for the model conditional on
  value \eqn{0} for the target binary variable value).
  \item \code{'1'} (or \code{1} or \code{TRUE}) for the model conditional on
  value \eqn{1} for the target binary variable value).
}}

\item{suffix}{parameter for the name of the edited version of the variable
under analysis (see 'Details').}

\item{...}{extra arguments to be used to form the default control argument.}
}
\value{
A \linkS4class{data.table} with two colums named with the LHS
variable of \code{formula} and \code{fit}. The first column contains every
single value of this variable in the input data set \code{data}. The second
column contains a \code{glm} object with the fit of the model for each
variable value.
}
\description{
This function is basically a wrapper for function
\code{\link{robustlogrm}} to fit a logistic regression model on each value
of an input variable of the data set.
}
\details{
Regarding \code{formula} see \code{\link{robustlogrm}}. Although the
target variables for models with \code{cond=''}, on the one hand, and with
\code{cond=0,1}, on the other hand, are different (edited in the first case
and raw in the latter), it is not necessary to specify the suffix in the
formula.

The input data set \code{data} must contain at least the following colums:
\enumerate{

\item \code{id.vars}: columns with the identification variables of each unit.

\item raw target variable: column with the raw values of the variable under
analysis. This is the LHS variable of \code{formula}.

\item edited target variable: column with the edited values of the variable
under analysis. The name must be the same as that of the raw target variable
with a suffix specified in the parameter \code{suffix}.

\item regressors: columns corresponding to the variables specified in the RHS
of \code{formula}.
}
}
\examples{
\dontrun{
regfit2(formula= 'y ~ x1', data= trainDT, id.vars= 'ID', cond= '', suffix= '')
}

}
