% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFfit.R
\name{RFfit}
\alias{RFfit}
\title{Fit random forest model}
\usage{
RFfit(data, targetVarName, parms, regressors)
}
\arguments{
\item{data}{data frame, list or environment (or object coercible by
\code{as.data.table} to a \linkS4class{data.table}) containing the variables
in the model (see 'Details').}

\item{targetVarName}{value of the regressand (an object of class character).}

\item{parms}{names of the identification variables for every unit in the
input data set \code{data}.}

\item{regressors}{an object of class character containing the design weight.}
}
\value{
an object of class randomForest with model fit.
}
\description{
\code{RFfit} fits a random forest model for a specific value of the regressand and 
specific regressors. Return an object of class randomForest.
}
\examples{

\dontrun{
  fitPar <- new(Class = 'fitParam',
  edData = FFall_AS.StQ, rawData = FGall_AS.StQ, 
  selParam = list(ntreeTry=1000, stepFactor=2, improve=0.05, 
                  trace=TRUE, plot=TRUE, doBest = TRUE, 
                  ptrain = 0.8, DD = DDactu),
                  valParam = list(edEffInd = effInd, priorBin = 5, 
                  dataVal = c('Train','Test')))
                  
 ObsPredPar1 <- new(Class = 'categObsPredModelParam',
                  Data = FGall_AS.StQ,
                  VarRoles = list(Units = IDUnits,
                  Domains = character(0),
                  DesignW = DesignW,
                  Regressands = Regressands,
                  Regressors = Regressors
                  ))
                  
ObsPredPar1 <-  fitModels(ObsPredPar1, fitPar, na.as.category)
}

}
