% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\docType{methods}
\name{setData<-,categObsPredModelParam,StQ-method}
\alias{setData<-,categObsPredModelParam,StQ-method}
\title{Set value of slot \code{Data} of an object \linkS4class{categObsPredModelParam}}
\usage{
\S4method{setData}{categObsPredModelParam,StQ}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Data} is to be assigned.}

\item{value}{Object of class \code{StQ} to be assigned to the slot \code{Data}.}
}
\value{
Object \linkS4class{categObsPredModelParam} with slot \code{Data} updated.
}
\description{
\code{setData} assigns a \code{StQ} to the slot \code{Data} of the input
object \linkS4class{categObsPredModelParam}
}

