% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setModelFits.R
\name{setModelFits<-}
\alias{setModelFits<-}
\title{Set value of slot \code{modelFits} of an object \linkS4class{categObsPredModelParam}}
\usage{
setModelFits(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{modelFits} is to be assigned.}

\item{value}{Object of class \code{data.table} to be assigned to the slot \code{modelFits}.}
}
\value{
Object \linkS4class{categObsPredModelParam} with slot \code{modelFits} updated.
}
\description{
\code{setModelFits} assigns a \linkS4class{data.table} to the slot \code{modelFits}
of the input object \linkS4class{categObsPredModelParam} with the fitted models for the
probabilities in the categorical observation-prediction model.
}

