% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setProbs.R
\name{setProbs<-}
\alias{setProbs<-}
\alias{setProbs<-,categObsPredModelParam,data.table-method}
\title{Set value of slot \code{probs} of an object \linkS4class{categObsPredModelParam}}
\usage{
setProbs(object) <- value

\S4method{setProbs}{categObsPredModelParam,data.table}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{probs} is to be assigned.}

\item{value}{\linkS4class{data.table} to be assigned to the slot \code{probs}.}
}
\value{
Object \linkS4class{categObsPredModelParam} with slot \code{probs} updated.
}
\description{
\code{setProbs} assigns a \linkS4class{data.table} to the slot \code{probs} of the
input object \linkS4class{categObsPredModelParam}.
}
