% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRegressands.R
\name{setRegressands<-}
\alias{setRegressands<-}
\alias{setRegressands<-,categObsPredModelParam,character-method}
\title{Set value of component Regressands of an object \linkS4class{categObsPredModelParam}}
\usage{
setRegressands(object) <- value

\S4method{setRegressands}{categObsPredModelParam,character}(object) <- value
}
\arguments{
\item{object}{Object \linkS4class{categObsPredModelParam} whose component \code{Regressands}
is to be assigned.}

\item{value}{character vector to be assigned to the component \code{Regressands}.}
}
\value{
Object \linkS4class{categObsPredModelParam} with the component \code{Regressands}
updated.
}
\description{
\code{setRegressands} assigns a character value to the component \code{Regressands}
of the slot \code{VarRoles} of the input object \linkS4class{categObsPredModelParam}.
}
