% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTest.R
\name{setTest<-}
\alias{setTest<-}
\alias{setTest<-,categObsPredModelParam,list-method}
\title{Set value of slot \code{Test} of an object \linkS4class{categObsPredModelParam}}
\usage{
setTest(object) <- value

\S4method{setTest}{categObsPredModelParam,list}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Test} is to be assigned.}

\item{value}{Object of class \code{list} to be assigned to the slot \code{Test}.}
}
\value{
Object \linkS4class{categObsPredModelParam} with slot \code{Test} updated.
}
\description{
\code{setTest} assigns a \code{list} to the slot \code{Test} of the input
object \linkS4class{categObsPredModelParam}
}
