% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTrain.R
\name{setTrain<-}
\alias{setTrain<-}
\alias{setTrain<-,categObsPredModelParam,list-method}
\title{Set value of slot \code{Train} of an object \linkS4class{categObsPredModelParam}}
\usage{
setTrain(object) <- value

\S4method{setTrain}{categObsPredModelParam,list}(object) <- value
}
\arguments{
\item{object}{Object whose slot \code{Train} is to be assigned.}

\item{value}{Object of class \code{list} to be assigned to the slot \code{Train}.}
}
\value{
Object \linkS4class{categObsPredModelParam} with slot \code{Train} updated.
}
\description{
\code{setTrain} assigns a \code{list} to the slot \code{Train} of the input
object \linkS4class{categObsPredModelParam}
}
