% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryingNewRegressorByVal.R
\name{tryingNewRegressorByVal}
\alias{tryingNewRegressorByVal}
\title{Returns a \linkS4class{list} with the global efficiency indicator and model fits for
prior probabilities for the best model selected.}
\usage{
tryingNewRegressorByVal(
  regformula,
  listRegressors,
  data,
  targetValue,
  id.vars,
  designWeight,
  parms,
  suffix
)
}
\arguments{
\item{regformula}{an object of class \code{\link[stats]{formula}} (or one that
can be coerced to that class): a symbolic description of the model to be fitted.}

\item{listRegressors}{an object of class character containing the name of all regressors
to possibly be considered in the model fit.}

\item{data}{data frame, list or environment (or object coercible by
\code{as.data.table} to a \linkS4class{data.table}) containing the variables
in the model (see 'Details').}

\item{targetValue}{value of the regressand (an object of class character).}

\item{id.vars}{names of the identification variables for each unit in the
input data set \code{data}.}

\item{designWeight}{an object of class character containing the design weight.}

\item{parms}{input parameters required for the process:
"maxit" for model fits,
"edEffInd" is a function used to define the global efficiency indicator,
"priorBin" contains the number of units in which the data is divided to construct 
the mesh of points to calculate the efficiency indicator.}

\item{suffix}{parameter for the name of the edited version of the variable under analysis.}
}
\value{
\linkS4class{list} containing three elements: 
The first one is a \linkS4class{list} with models fits for prior probabilities, description of 
the model to be fitted, value of the regressand and global efficiency indicator obtained 
for this specific value of the regressand.
The second one contains a symbolic description of the best model chosen among all models tried.
Last one element contains all models tried and the global efficiency indicator for each one.
}
\description{
\code{tryingNewRegressorByVal} computes all possible model fits for 
\code{regformula}, taking into account a new additional regressor among those 
existing in \code{listRegressors}. 
Also, \code{tryingNewRegressorByVal} compares all fits and chooses the best model,
according to the global efficiency indicator. For the selected model, model fits for
prior probabilities is returned.
}
\examples{

\dontrun{
}

}
