% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitParam-class.R
\docType{class}
\name{fitParam-class}
\alias{fitParam-class}
\title{S4 class for the parameters necessary to fit a categorical observation-prediction model}
\description{
Definition of the S4 class named \code{fitParam} for the parameters needed to fit
a categorical observation-prediction model in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{edData}}{object of class \code{\link[StQ]{StQ}} with the edited and validated values.}

\item{\code{rawData}}{object of class \code{\link[StQ]{StQ}} with the raw values.}

\item{\code{selParam}}{list of parameters for the model selection when \code{selection == TRUE}}

\item{\code{valParam}}{list of parameters for validation.}
}}

\examples{
# An empty contObsPredModelParam object:
new(Class = 'fitParam')

\dontrun{
source('R:/USIE/Proyecto_DepSel_VarQual/datos/PreparacionStQsAnalisis.R', encoding = 'UTF-8')
fitPar <- new(Class = 'fitParam', edData = ff_2011_model.StQ, rawData = fd_2011_model.StQ,
             selParam = list())

}

}
