% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarRoles.R
\docType{methods}
\name{getVarRoles}
\alias{getVarRoles}
\alias{getVarRoles,categObsPredModelParam-method}
\title{Return slot \code{VarRoles} from the input object}
\usage{
getVarRoles(object)

\S4method{getVarRoles}{categObsPredModelParam}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{categObsPredModelParam}.}
}
\value{
List with components \code{Units}, \code{Domains}, \code{DesignW}, \code{Regressands},
\code{Regressors}.
}
\description{
\code{getVarRoles} extracts the slot \code{VarRoles} of the input object.
}
\examples{

}

